Modularizing the chat api & refactoring the code
