+++ chatapi/src/config/openai.config.ts
@@ -1,0 +1,12 @@
+import { Configuration, OpenAIApi } from 'openai';
+import dotenv from 'dotenv';
+
+dotenv.config();
+
+const configuration = new Configuration({
+  'apiKey': process.env.OPENAI_API_KEY,
+});
+
+const openai = new OpenAIApi(configuration);
+
+export default openai;
